+++ docs/.vuepress/config/en/series.ts
@@ -1,0 +1,15 @@
+export const series = {
+  // todo README.md 无法展示，没有对 '' 进行处理
+  '/en/docs/guide/': ['introduce', 'architecture', 'getting-started'],
+  '/en/docs/theme/': [
+    {
+      text: 'Basic',
+      children: ['introduce', 'usage', 'custom-catalog-title']
+    },
+    {
+      text: 'Advance',
+      children: ['home', 'series', 'comments', 'auto-set-category', 'custom-container', 'code-import', 'bulletin-popover', 'register-components', 'password']
+    }
+  ],
+  '/en/docs/plugins/': ['page', 'comments', 'vue-previews', 'bulletin-popover'],
+}