+++ packages/vuepress-theme-reco/CHANGELOG.md
@@ -1,5 +1,47 @@
 # vuepress-theme-reco
 
+## 2.0.0-beta.29
+
+### Patch Changes
+
+- fix: fix some error
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.29
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.29
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.29
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.29
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.29
+  - @vuepress-reco/shared@2.0.0-beta.29
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.29
+
+## 2.0.0-beta.28
+
+### Patch Changes
+
+- fix: optimize the style
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.28
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.28
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.28
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.28
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.28
+  - @vuepress-reco/shared@2.0.0-beta.28
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.28
+
+## 2.0.0-beta.27
+
+### Patch Changes
+
+- fix: optmize some error
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.27
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.27
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.27
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.27
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.27
+  - @vuepress-reco/shared@2.0.0-beta.27
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.27
+
 ## 2.0.0-beta.26
 
 ### Patch Changes